Fixed problems with incorrect font sizes in CSS Magic. CSS Magic now recognises more stylesheets, icluding @import commands.
Template Voodoo does not yet do anything for forward integrations -- the only way to make it work will be to stop using the phpBB template system, or to get CSS Magic to make changes to CSS *first*, before generating the page.

The latter is probably the better solution as it will likely also improve page responsiveness. However, for now, CSS Magic is doing enough to improve CSS conflicts on integrated pages.

git-svn-id: http://www.wp-united.com/svn/trunk@144 c1c82642-cce3-494f-bf0e-9d5eb4df7846